Rename src/open_data_pvnet/scripts to scripts/generation
